Revert dexmaker cache property after tests
Also moved the base test case class to a more appropriate Java package
